+++ spec/ig_markets/cli/commands/activities_command_spec.rb
@@ -21,13 +21,13 @@ describe IGMarkets::CLI::Main, :cli_command do
   end
 
   it 'prints activities from a number of days and a from time' do
-    from = Time.new 2015, 1, 15, 6, 0, 0
-    to = Time.new 2015, 1, 18, 6, 0, 0
+    from = Time.new 2015, 1, 15, 6, 0, 0, '+04:00'
+    to = Time.new 2015, 1, 18, 6, 0, 0, '+04:00'
 
     expect(dealing_platform.account).to receive(:activities).with(from: from, to: to).and_return([])
 
     expect do
-      cli(days: 3, from: '2015-01-15T06:00:00', sort_by: 'date').activities
+      cli(days: 3, from: '2015-01-15T06:00:00+04:00', sort_by: 'date').activities
     end.to output("#{IGMarkets::CLI::Tables::ActivitiesTable.new []}\n").to_stdout
   end