+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 0.32 — Unreleased
 
+- All timestamps provided to the `ig_markets activities`, `ig_markets transactions`, and `ig_markets performance`
+  commands must now specify their time zone explicitly
 - Support the `unavailable` state on `IGMarkets::Streaming::MarketUpdate#market_state`
 
 ### 0.31 — March 26, 2017